requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
requests==2.32.3
python-bitcoinlib==0.11.2
pydantic==2.7.1
sqlalchemy==2.0.30
asyncpg==0.29.0
motor==3.4.0
bcrypt==4.1.3
PyJWT==2.8.0
GitPython==3.1.43
bitcoinlib==0.7.3
aiosqlite
